Backend-Task-page/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
certifi==2025.8.3
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
dnspython==1.16.0
ecdsa==0.19.1
et_xmlfile==2.0.0
fastapi==0.115.12
h11==0.14.0
idna==3.10
numpy==2.3.2
openpyxl==3.1.5
pandas==2.3.2
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pymongo==3.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.4.0
pytz==2025.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.34.2
xlsxwriter==3.2.9
